requirements.txt
numpy
scipy
docopt
pygments
python-Levenshtein
# matplotlib  # for option -p
